+++ NEWS.md
@@ -1,6 +1,7 @@
 # Unreleased version
 
 - New feature: you can now reset all inputs on the page by calling `reset()` with no arguments (#222)
+- New feature: Add a `removeEvent()` function which removes events added to HTML elements with `onclick()` or `onevent()` (#244)
 - Fix bug: `disable()` did not work on nested download buttons (#223)
 - Fix bug: Don't automatically namespace ID arguments in custom extendShinyjs functions (#229)
 - Fix bug: ensure that `extendShinyjs()` functions don't overwrite native {shinyjs} functions (#230)